Scrape book reviews from Goodreads using RSelenium

This function scrapes book reviews from Goodreads based on provided book IDs using RSelenium for dynamic content loading to retrieve more than 30 reviews.

Usage

scrape_reviews(
  book_ids_path,
  num_reviews = 30,
  use_parallel = FALSE,
  num_cores = 4,
  delay_min = 2,
  delay_max = 5,
  browser_type = "firefox",
  headless = TRUE
)

Arguments

book_ids_path

A character string specifying the path to a file containing book IDs.

num_reviews

An integer specifying the number of reviews to scrape per book. Default is 30.

use_parallel

A logical value indicating whether to use parallel processing. Default is FALSE.

num_cores

An integer specifying the number of cores to use for parallel processing. Default is 4.

delay_min

Minimum delay between requests in seconds. Default is 2.

delay_max

Maximum delay between requests in seconds. Default is 5.

browser_type

Browser type for RSelenium ("firefox", "chrome", "phantomjs"). Default is "firefox".

headless

Whether to run browser in headless mode. Default is TRUE.

Value

A data frame containing scraped review information.

Examples

# \donttest{
# Create a temporary file with sample book IDs
temp_file <- tempfile(fileext = ".txt")
writeLines(c("1420", "2767052", "10210"), temp_file)
# Run the function (with a small number of reviews to keep the example quick)
reviews <- scrape_reviews(temp_file, num_reviews = 50, use_parallel = FALSE)
#> Total book IDs to process: 3
#> Target reviews per book: 50
#> Using sequential processing...
#> Starting to scrape reviews for book ID: 1420
#> 
#> Selenium message:Process unexpectedly closed with status 1
#> Build info: version: '4.0.0-alpha-2', revision: 'f148142cf8', time: '2019-07-01T21:30:10'
#> System info: host: 'pkrvmbietmlfzoi', ip: '10.1.0.145', os.name: 'Linux', os.arch: 'amd64', os.version: '6.11.0-1015-azure', java.version: '17.0.15'
#> Driver info: driver.version: unknown
#> remote stacktrace: 
#> Could not open firefox browser.
#> Client error message:
#> 	 Summary: UnknownError
#>  	 Detail: An unknown server-side error occurred while processing the command.
#> 	 Further Details: run errorDetails method
#> Check server log for further details.
#> Error processing book ID 1420: Failed to setup RSelenium driver: Undefined error in httr call. httr output: length(url) == 1 is not TRUE
#> Starting to scrape reviews for book ID: 2767052
#> 
#> Selenium message:Process unexpectedly closed with status 1
#> Build info: version: '4.0.0-alpha-2', revision: 'f148142cf8', time: '2019-07-01T21:30:10'
#> System info: host: 'pkrvmbietmlfzoi', ip: '10.1.0.145', os.name: 'Linux', os.arch: 'amd64', os.version: '6.11.0-1015-azure', java.version: '17.0.15'
#> Driver info: driver.version: unknown
#> remote stacktrace: 
#> Could not open firefox browser.
#> Client error message:
#> 	 Summary: UnknownError
#>  	 Detail: An unknown server-side error occurred while processing the command.
#> 	 Further Details: run errorDetails method
#> Check server log for further details.
#> Error processing book ID 2767052: Failed to setup RSelenium driver: Undefined error in httr call. httr output: length(url) == 1 is not TRUE
#> Starting to scrape reviews for book ID: 10210
#> 
#> Selenium message:Process unexpectedly closed with status 1
#> Build info: version: '4.0.0-alpha-2', revision: 'f148142cf8', time: '2019-07-01T21:30:10'
#> System info: host: 'pkrvmbietmlfzoi', ip: '10.1.0.145', os.name: 'Linux', os.arch: 'amd64', os.version: '6.11.0-1015-azure', java.version: '17.0.15'
#> Driver info: driver.version: unknown
#> remote stacktrace: 
#> Could not open firefox browser.
#> Client error message:
#> 	 Summary: UnknownError
#>  	 Detail: An unknown server-side error occurred while processing the command.
#> 	 Further Details: run errorDetails method
#> Check server log for further details.
#> Error processing book ID 10210: Failed to setup RSelenium driver: Undefined error in httr call. httr output: length(url) == 1 is not TRUE
#> 2025-06-29 02:32:07.998916 scrape_goodreads_reviews: Completed! All book reviews extracted
#> Scraping run time = 5.62692189216614
#> Total books processed: 3
#> Total reviews collected: 3
#> Average reviews per book: 1.0
print(head(reviews))
#> # A tibble: 3 × 8
#>   book_id reviewer_id reviewer_name review_content reviewer_followers
#>   <chr>   <lgl>       <lgl>         <lgl>          <lgl>             
#> 1 1420    NA          NA            NA             NA                
#> 2 2767052 NA          NA            NA             NA                
#> 3 10210   NA          NA            NA             NA                
#> # ℹ 3 more variables: reviewer_total_reviews <lgl>, review_date <lgl>,
#> #   review_rating <lgl>
# Clean up: remove the temporary file
file.remove(temp_file)
#> [1] TRUE
# }